Restructure cr.cube tests by specific functionality
